fixed api; reformatted code
