--- custom_components/CryptoTrakcer/translations/it.json
@@ -1,21 +1,0 @@
-{
-    "config": {
-        "title": "Cryptotracker",
-        "step": {
-            "user": {
-                "title": "CryptoTracker",
-                "description": "If you need help with the configuration have a look here: https://github.com/custom-components/integration_blueprint",
-                "data": {
-                    "name": "Name",
-                    "compare": "Compare"
-                }
-            }
-        },
-        "error": {
-            "invalid_format": "Formato invalido"
-        },
-        "abort": {
-            "single_instance_allowed": "Only a single instance is allowed."
-        }
-    }
-}
