+++ custom_components/cryptostate/manifest.json
@@ -8,5 +8,5 @@
     "requirements": [],
     "config_flow": true,
     "iot_class": "local_polling",
-    "version": "1.1.2"
+    "version": "2.0.0"
 }
