+++ custom_components/cryptostate/translations/it.json
@@ -1,0 +1,19 @@
+{
+    "config": {
+        "title": "Cryptotracker",
+        "step": {
+            "user": {
+                "title": "CryptoTracker",
+                "description": "Se hai bisogno di aiuto, vai alla pagina del componente: https://github.com/BigNocciolino/CryptoTracker",
+                "data": {
+                    "name": "Nome",
+                    "crypto": "Crypto",
+                    "base": "Base"
+                }
+            }
+        },
+        "error": {
+            "invalid_format": "Formato invalido, per una lista di monete supportate guarda: https://cdn.jsdelivr.net/gh/fawazahmed0/currency-api@1/latest/currencies.json"
+        }
+    }
+}
